+++ bdld/actions/__init__.py
@@ -1,3 +1,5 @@
+"""Import all submodules when importing the actions module"""
+# when doing `import bdld.actions`
 from . import action
 from . import birth_death
 from . import bussi_parinello_ld
@@ -8,6 +10,7 @@ from . import overdamped_ld
 from . import particle_distribution
 from . import trajectory_action
 
+# when doing `from bdld.actions import *` (not recommended)
 __all__ = [
     "action",
     "birth_death",